attempt to remove file before copying another

+++ src/commands/makeMigration.ts
@@ -100,7 +100,9 @@ const command: MakeMigrationCommand = async (name: string, blank = false): Promi
 
         // copy current schema for future reverts
         const currentSchemaBackup = path.join(migrationsPath, newMigration, SCHEMA_FILE_NAME)
-        fs.rmSync(currentSchemaBackup)
+        if (fs.existsSync(currentSchemaBackup)) {
+            fs.rmSync(currentSchemaBackup)
+        }
         fs.copyFileSync(schema, currentSchemaBackup)
 
         // create a revert migration script based on previous schema
@@ -118,7 +120,9 @@ const command: MakeMigrationCommand = async (name: string, blank = false): Promi
 
             // cleanup
             fs.rmSync(path.join(migrationsPath, revertMigration), { recursive: true })
-            fs.rmSync(schema)
+            if (fs.existsSync(schema)) {
+                fs.rmSync(schema)
+            }
             fs.copyFileSync(currentSchemaBackup, schema)
         }
